Support arm7l, aarch64 architectures and manylinux2014 tag
The architecture, wheel names, and shared object names
match what is specified in CPython and pip.

arm7l and aarch64 are supported for Linux only.

Target::from_target_triple() returns an Err on an
unsupported OS and architecture combination.

This allows building arm7l and aarch64 on manylinux1 and
manylinux2010, which should not work in pip.
Target::from_target_triple() could be expanded to error
on this but that needs more refactoring than seems
appropriate here.