Add a workaround for the GitLab PyPI API implementation
GitLab includes its own PyPI upload API implementation
that is not based on Warehouse.
It insists that "requires_python" metadata field must
always be sent, even when it's empty.

The package upload to GitLab PyPI fails with:

    400 Bad Request: {"error":"requires_python is missing"}

when it's not included in the request.

The official API reference does not mandate this, though:

    https://warehouse.readthedocs.io/api-reference/legacy.html

Unfortunately, twine always includes these metadata fields
in its requests, even when they are empty, and the GitLab
API implementation happened to depend on this.

Add a workaround that mimics twine behavior and silently
includes an empty "requires_python" field even when
the package developers do not define "Requires-Python".