Add `extra-src` to data
This adds a custom `extra-src` subdirectory to the .data folder, which gets installed in the site-packages.

I'm not completely sure if that addition is a good idea, but i unfortunately need to add a `.pth` and don't have a better idea. it is also similarly dangerous to the existing `.data/data`