Build platform independent rust binary wheel through wasi (#1107)
* Build platform independent rust binary wheel through wasi

We build the rust binaries to wasi and put them in a wheel with a wasmtime launcher. The resulting wheels are platform independent (py3-none-any) and run on all wasmtime platforms. Those are currently less than maturin natively, but you still only have to build one wheel and i hope for some nice things in the future (isolation capabilities, wasm runtimes on less-supported platforms)

* Test fixes

* fix typo

* Add wasm32-wasi target

* Add wasm32-wasi target in cirrus ci

* windows ci

* bin name validation and wasmtime 1.0

* changelog