Use external `uniffi-bindgen` if no root package is configured
A "root package" in Cargo exists in a workspace if the workspace
Cargo.toml also has a `[package]` section.
In that case maturin tries to find a binary target called
`uniffi-bindgen` to use.

But if there's no root package to begin with it should fall back to an
external `uniffi-bindgen` instead of panicking.