Fix platform tags when cross-compiling universal2 (#2153)
When cross-compiling for `universal2-apple-darwin`, the platform tags of
the resulting wheel were not properly derived. This is because with
forcing `target_triple` to `None` as before, it was not detected that we
are in fact cross-compiling for a different target and thus the wheels
were wrongly tagged for the host platform.

To make the inference work, set `target_triple` to a valid triple
instead.