Add current package context to source dist error
Place source dist readmes next to Cargo.toml

`package.readme` may point to any point above the package, so when we move the directory, but keep the readme position, we could get different readme files at the same archive location. To solves this, we rewrite `Cargo.toml` to find the readme in the same directory and add the readme to this directory.