Forward `cargo package --list` warnings (#2186)
Forward warnings from `cargo package --list` calls. Motivation is the `crates/jiter` crate in [jiter](https://github.com/pydantic/jiter):

```
$ cargo package --list
warning: readme `../../README.md` appears to be a path outside of the package, but there is already a file named `README.md` in the root of the package. The archived crate will contain the copy in the root of the package. Update the readme to point to the path relative to the root of the package to remove this warning.
```