Bump metadata version to 2.4 to fix license bug
When a project sets `license = { file = "LICENSE" }`, we're currently emitting a `License-File` entry, but Metadata-Version 2.3, which is invalid and now gets rejected by PyPI. This change bumps the metadata version.

Alternatively, we could only bump the version if license files are used and use the old `License` field for `license = { file = "LICENSE" }` again, but i'm not aware of any index currently rejecting metadata version 2.4.