Remove hard coded `pip show` in `fix_direct_url`
Before uv 0.4.25, `uv pip show` did not support the `--files` argument
which is required by `fix_direct_url`. Now that the flag is supported,
falling back to pip is not necessary.

Error handling is also improved. Previously if pip was not present
in the environment (which is the case for `uv venv`) no error was
reported to the user but `fix_direct_url` will have failed so the
package is not installed in editable mode.