check the install backend version to give more informative error messages (#2389)
since version 0.4.25, `uv pip show --files` is supported and can be used
to set the package installed by `maturin develop` as editable without
having to fallback to using `pip`. However several users are using old
versions of `uv` and it would be better if the error message was clearer
if an outdated uv is used.

This has caused confusion in these issues:
- https://github.com/PyO3/maturin/discussions/2388
- https://github.com/PyO3/maturin-import-hook/issues/15

before:
```
✏️  Setting installed package as editable
💥 maturin failed
  Caused by: failed to list package files
```
after:
```
✏️  Setting installed package as editable
⚠️ Warning: failed to set package as editable: uv >= 0.4.25 is required for `show --files`. Version 0.4.23 was found.
🛠 Installed $PACKAGE
```
(note that installation now proceeds if setting as editable fails)