Make `maturin develop` fail if version info is missing in `pyproject.toml` (#2418)
This PR implements the changes discussed in
https://github.com/PyO3/maturin/issues/2416#issuecomment-2565278865 on
top of #2417. Got this done faster than expected this morning, but won't
have any time to incorporate feedback for the next ~ two days. Feel free
to pull this in and butcher it any way you like. :)

Thanks again for your fantastic work on maturin! 🍰

---------

Co-authored-by: messense <messense@icloud.com>