Handle archived dylibs on AIX (#2442)
In recent Rust versions, the dylib format for AIX has changed to reflect
that shared libraries on AIX are normal archived, in the Big archive
format.
See https://github.com/rust-lang/rust/pull/132362.

However, Python wheels on the platforms still use un-archived shared
objects, so this change updates module_writer to unarchive the archived
dylib before we copy it over to the wheel.