Fix platform tag on Solaris/Illumos (#2483)
This fixes platform tag on Solaris/Illumos.

The problem is that `os` here can never be `sunos`, but rather
`solaris`/`illumos`.

The following:
```diff
-                if os.starts_with("sunos") {
+                if os.starts_with("illumos") || os.starts_with("solaris") {
```
works as well, but checking the enum values seems nicer (and it's
definitely faster). Using `get_python_os` would likely work as well, but
I didn't go that way for the same reason.