fix auditwheel .so relocation for namespace modules (#2513)
For a project of mine I have a setup as a namespace package like this:

``` toml
[tool.maturin]
python-source = "python/src"
module-name = "my.example.namespace.module"
```

The module has a shared library dependency and currently the shared
object generated by audit wheel fails to load as the rpath gets set
incorrectly. The wheel has the following contents

```
/my.libs/my_dep.so
/my/example/namespace/module.so
```
The `RATH`/`RUNPATH` for `/my/example/namespace/module.so` gets set to
`$ORIGIN/../my.libs/my_dep.so` which means that `my_dep.so` is not found
while trying to load the module. Instead the `RATH`/`RUNPATH` should be:
`$ORIGIN/../../../my.libs/my_dep.so`.

I took a stab at fixing this. It looks like the logic for determining
the artifact path was wrong and assumed the shared library would endup
in `my.example.namespace.module/module.so` but it endsup in
`my/example/namespace/module.so`.