Release v1.8.3 (#2515)
As suggested I wrote up a changelog and bumped versions to help with the
release.

I tried to mimic the style/level of detail of the changelog. I am not
super familiar with the codebase so I may have not included something
that should have been there. I appreciate the quick turnaround!