Fix broken `maturing develop` with latest uv
This patch fixes `maturin develop` not working with the latest versions
of uv due to breaking changes in uv. Previously the `--version` argument
could be supplied to any subcommand, but since v0.7 that argument can
only be supplied as `uv --version`.

Specifically, the fix changes the `uv pip` command to `uv` when we get
the uv version as part of the check if `show --files` is supported.