Fix broken `maturin develop` with latest uv (#2584)
This patch fixes `maturin develop` not working with the latest versions
of uv due to breaking changes in uv. Previously the `--version` argument
could be supplied to any subcommand, but since v0.7 that argument can
only be supplied as `uv --version`.

Specifically, the fix changes the `uv pip` command to `uv` when we get
the uv version as part of the check if `show --files` is supported.

Without this patch `configure_as_editable` fails due to
`check_supports_show_files` failing, which again fails because
`InstallBackend::version` fails with uv 0.7.x or newer.

See https://github.com/astral-sh/uv/pull/13108 for more details.

I've not added any new tests since I believe the current test suite
covers this change. The change is backwards compatible with older
versions of uv. But if a new test is required, please advice exactly
what/how you want tested, and I'll add it.