Self contained rust boostrapping (#2421)
Currently, maturin is not self-contained: You need to install rust on
the machine in order to build a package. We can improve this by
downloading a matching rustup, using it to install cargo and rustc into
a temporary directory and using this rust for the build. As long as the
platform is supported by rustup, we get a self-contained build that
doesn't mess with the host system. You can for example use `pip install
...` in a fresh `python3-slim` docker container and any rust dependency
with a missing wheel will build and install without additional setup
(while it currently fails and prompts you to install rust).

To inject this cache installation, we need to set `PATH`, `RUST_HOME`
and `CARGO_HOME` whenever we call `cargo` or `rustc`.

Since setuptools rejects supporting programmatically setting
dependencies (https://github.com/pypa/setuptools/issues/2854), we use
their hack of an in-tree build backend that overrides some of
setuptools' hooks. In this hook, we install
https://github.com/konstin/puccinialin if rust is missing, and later (in
`setup.py`) run it to install rust and run maturin with the right env
vars already set.

There's an option to disable rust bootstrapping by setting
`MATURIN_NO_INSTALL_RUST` for users that don't want this automatism
(e.g. linux distros).

In effect, this now works (using the dev maturin for demonstration):

```
docker run --rm -it -v $(pwd):/io python:3.13 python -m pip install /io/ -v
```

I've only tested linux/windows/mac and x86_64/arm64 in puccinialin so
far, but i hope to add all the other rustup supporter platforms and
arches over time.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>