bootstrap: fix overriden setuptools PEP 517 hooks (#2612)
config_settings is not supposed to have an underscore as per PEP 517.
This changes the argument name to adhere to the PEP. Currently, this
breaks pyproject-rpm-macros (Python build frontend used by Fedora Linux)
which tries to pass a kwarg named `config_settings` and blows up. (We
are considering changing pyproject-rpm-macros to pass it as a positional
argument like pyproject-hooks/pip does, but this fix is still valid.)

Additionally, the overriden get_requires_for_* hooks still should
include the values from setuptools. Older setuptools versions include
"wheel" here, for example.