Allow specifying compression method and level, in both `build` and `develop` modes. (#2625)
This is a follow-up to https://github.com/PyO3/maturin/issues/2566 that
adds the ability to
- Specify other compression methods
- Control compression in the `develop` mode, not only `build`. Solves
#2622

```
--compression-method <COMPRESSION_METHOD>
    Zip compresson method. Only Stored and Deflated are currently compatible with all package managers

    [default: deflated]

    Possible values:
    - deflated: Deflate compression
    - stored:   No compression
    - bzip2:    BZIP2 compression
    - zstd:     Zstandard compression (supported from Python 3.14)

--compression-level <COMPRESSION_LEVEL>
    Zip compresson level

    [default: 6]
```

Note that additional compression methods were added in `uv` 0.7.3 in
https://github.com/astral-sh/uv/pull/13285 but adding local `whl` is
broken until https://github.com/astral-sh/uv/pull/13781 is merged.

I tested the change locally by setting
```
        compression: maturin::CompressionOptions {
            compression_method: maturin::CompressionMethod::Bzip2,
            compression_level: 6,
        },
```
in the tests, which is supported by `pip`. I can do a follow-up PR to
update the tests after `uv` is updated, possibly with a version check.