make `PythonInterpreter::check_executable` aware of windows arch (#2758)
Fixes #2692 (try 2)

This PR proposes removing the checks in `windows_interpreter_no_build`
into `PythonInterpreter::check_executable`, making that the single way
to resolve if an interpreter has all the information to build.

We need to run that step anyway, so integrating the windows checks into
it should make things more efficient.

At the same time, I also made `check_executable` aware of architecture
when falling back to `py` to search for an interpreter.