Add support for iOS cross-platform virtual environments (#2828)
Tools such as `cibuildwheel` and
[`xbuild`](https://pypi.org/project/xbuild/) manage building wheels for
iOS by using a cross-platform virtual environment - a valid virtual
environment for the build platform that can "pretend" to be the
target/host platform when asked for sysconfigdata. This is done by
monkeypatching key values in the sys, sysconfig, platform and os module
when the interpreter starts. Many of these environments also set
`sys.cross_compiling = True` as an indicator that they aren't "real"
environments.

Maturin's concept of "cross-compiling", however, has more to do with
whether the interpreter that is running the Maturin build can be used to
interrogate details about the Python install for the purpose of
configuring a build. To that end, somewhat counterintuitively, a
cross-platform venv *isn't* "cross platform" from the perspective of
Maturin.

This PR makes 3 changes:
1. It marks iOS builds as *not* being cross-platform. This ensures that
the sysconfigdata values from the interpreter running maturin are passed
to the build
2. When the build has not been given an explicit `--target`, or
`--interpreter`, the build interpreter will be interrogated to see if it
is a cross-compilation environment - and if it is, the target implied by
the platform is used for the build.
3. When an explicit `--interpreter` *has* been provided, the target will
be modified if the environment is a cross-platform environment.

When combined with #2827, this means it is possible to create an iOS
wheel with no more than `maturin build`, provided you're in an iOS
cross-platform venv (such as those created by `cibuildwheel` and
`xbuild`).