Make sdist reproducible (#2831)
I compared an sdist uploaded to pypi to one built from source at the
same tag and found two cases of metadata making the build not
reproducible:
1. The uid/gid data is not cleared out (for both reproducibility and
security, these should be 0/root)
2. The file mtime seems to be the build timestamp instead of
`SOURCE_DATE_EPOCH` or some other fixed value

This PR combines 2 changes that both, individually, can address both
issues above:
1. Ensure that we use `HeaderMode::Deterministic` when creating the
sdist tar. This ensures that all files added using `add_file()` have
uid/gid set to 0 and an mtime of July 23, 2006
2. Remove the override of the `add_file()` method in the `SdistWriter`
implementation of `ModuleWriter`. This forces all writes to go through
`add_bytes_with_permissions()`, which never had an issue with the
uid/gid and, as of this PR, also sets the mtime to the static value from
the tar crate.

This means the fix 1 isn't strictly required as long as no one writes to
the archive without going through `add_bytes_with_permissions()` but, if
they do, those files should look identical to everything else in the
archive.

~~This PR does not honor `SOURCE_DATE_EPOCH` if it is set in the build
env, but it probably would not be a lot of work to add it.~~