Linux sw_64 architecture support
In this commit, support for a new architecture sw_64 is added to maturin. sw_64 is a new RISC ISA, which is a bit like RISC-V. Though it is a rather new ISA, it has been added to openEuler and openAnolis port since 2020. So supporting this new architecture is meaningful.
It's PR is to solve compilation failure on sw_64 architecture.

Binutils has identified EM_SW64 is 268, I hope maturin can accept this patch. Thanks.

https://en.wikipedia.org/wiki/Sunway_(processor)