Set entry type when adding to the tar file (#2859)
We need to explicitly set this because the header initializes with all
null bytes. When this is not set, the python `tarfile` module can fail
to process the file properly due to treating an entry of type `AREGTYPE`
that also ends in a trailing slash as a directory instead of a regular
file under some very specific circumstances.
See https://github.com/python/cpython/issues/141707 for more details

Fixes #2855