Migrate uniffi bindings to new `BindingGenerator` trait (#2878)
Unlike the other binding modes, uniffi actually generates files on the
filesystem that need to be included in the archive, not just in-memory
strings. I didn't want to read all those files into memory so my
solution was to add an `ArchiveSource` enum. That's the first commit in
this PR. The second commit actually adds the new generator impl.