Clean up BindingGenerator interface (#2881)
The last PR about the `BindingGenerator` trait! (at least for now 😅 )

Now that all the bindings have been migrated, it seems fairly clear that
the temp dir and python interpreter only matter for some bindings and
shouldn't be a part of the trait interface itself. This PR hoists those
parameters into the generator struct if necessary and adjusts the code
to match.

It also creates a `BindingType` enum for PyO3 so that we can error out
early if the python interpreter is required but missing, rather than
panicking later on.