Update ModuleWriter logic to use ArchiveSource enum (#2882)
This PR updates the `ModuleWriter` trait to use the `ArchiveSource` enum
that was added to support the `BindingGenerator` trait.

This is part 1 of 2 PRs, where the second one will add a `VirtualWriter`
struct that will wrap around `WheelWriter` and `SDistWriter`. The
virtual writer will store a `HashMap<PathBuf, ArchiveSource>` internally
so, with this update, the virtual writer can eventually just call
`self.inner.add_entry(target, source)`.

The primary purpose of this virtual writer is to ensure that the files
are sorted by path before being written to the underlying archive and
ensure that builds are reproducible. As a useful side effect, this will
move the file exclusion and duplicate tracking to the virtual writer as
well, providing a nice separation of concerns with the virtual writer
deciding if a file should be present in the archive and the underlying
writer actually doing the writing to the archive.

The intended end result for this set of patches can be viewed on [my
virtual-writer
branch](https://github.com/e-nomem/maturin/blob/virtual-writer/src/module_writer/virtual_writer.rs).