Auto-enable required features for `uniffi-bindgen` (#2886)
I'm not sure if the behavior here should be to plumb in the enabled
features from the higher levels of maturin to honor the explicitly
selected features from the user or to just automatically enable the
required features to run the `uniffi-bindgen` binary. I went with the
second option here mostly because it's easier.

Fixes #2806