Add `VirtualWriter` to track and order archive entries (#2887)
As I mentioned previously in https://github.com/PyO3/maturin/pull/2882,
this PR adds a `VirtualWriter` that wraps around the existing module
writer structs and extracts the common file tracking and exclusion
logic. In addition, it tracks all files before writing them at the end
allowing us to write the files with a consistent order.

After this PR, it should be possible to have mostly reproducible builds.
For sdists this requires using the same version of maturin for
reproducibility, and for wheels it requires using the same versions of
maturin, the rust toolchain, and external libraries.